requirements.txt
python-binance
argparse
